requirements.txt
selenium==4.25.0
gspread==6.1.4
oauth2client==4.1.3
google-auth==2.35.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
crawl4ai
playwright
